+++ src/main/java/sonia/scm/pushlog/PushedByContributorChangesetPreProcessorFactory.java
@@ -27,9 +27,9 @@ import com.google.common.base.Strings;
 import sonia.scm.plugin.Extension;
 import sonia.scm.repository.ChangesetPreProcessor;
 import sonia.scm.repository.ChangesetPreProcessorFactory;
+import sonia.scm.repository.Contributor;
 import sonia.scm.repository.Person;
 import sonia.scm.repository.Repository;
-import sonia.scm.repository.Trailer;
 import sonia.scm.user.DisplayUser;
 import sonia.scm.user.UserDisplayManager;
 
@@ -37,15 +37,15 @@ import javax.inject.Inject;
 import java.util.Optional;
 
 @Extension
-public class PushedByTrailerChangesetPreProcessorFactory implements ChangesetPreProcessorFactory {
+public class PushedByContributorChangesetPreProcessorFactory implements ChangesetPreProcessorFactory {
 
-  private static final String TRAILER_TYPE = "Pushed-by";
+  private static final String CONTRIBUTOR_TYPE = "Pushed-by";
 
   private final PushlogManager pushlogManager;
   private final UserDisplayManager userDisplayManager;
 
   @Inject
-  public PushedByTrailerChangesetPreProcessorFactory(PushlogManager pushlogManager, UserDisplayManager userDisplayManager) {
+  public PushedByContributorChangesetPreProcessorFactory(PushlogManager pushlogManager, UserDisplayManager userDisplayManager) {
     this.pushlogManager = pushlogManager;
     this.userDisplayManager = userDisplayManager;
   }
@@ -56,20 +56,20 @@ public class PushedByTrailerChangesetPreProcessorFactory implements ChangesetPre
       String pusherName = pushlogManager.get(repository).get(changeset.getId());
 
       if (!Strings.isNullOrEmpty(pusherName)) {
-        Trailer trailer = createTrailer(pusherName);
-        changeset.addTrailer(trailer);
+        Contributor contributor = createContributor(pusherName);
+        changeset.addContributor(contributor);
       }
     };
   }
 
-  private Trailer createTrailer(String pusherName) {
+  private Contributor createContributor(String pusherName) {
 
     Optional<DisplayUser> optionalDisplayUser = userDisplayManager.get(pusherName);
     if (optionalDisplayUser.isPresent()) {
       DisplayUser displayUser = optionalDisplayUser.get();
-      return new Trailer(TRAILER_TYPE, new Person(displayUser.getDisplayName(), displayUser.getMail()));
+      return new Contributor(CONTRIBUTOR_TYPE, new Person(displayUser.getDisplayName(), displayUser.getMail()));
     } else {
-      return new Trailer(TRAILER_TYPE, new Person(pusherName));
+      return new Contributor(CONTRIBUTOR_TYPE, new Person(pusherName));
     }
   }
 }